cleaned up layout.js, moved storm point code and state to active_storm_list.js component and storm_utils.js
